More C tests passing.
